# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NRWA API Documentation",
        "description": "REST API dokumentacija za NRWA projekt",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/v1/dispatchers": {
            "get": {
                "tags": [
                    "Dispatchers"
                ],
                "summary": "Get list of dispatchers",
                "operationId": "d193b74eccd768ead409d2eda396c439",
                "responses": {
                    "200": {
                        "description": "List of dispatchers"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dispatchers"
                ],
                "summary": "Create new dispatcher",
                "operationId": "6ddf4e52dd56f1dd6f4a5ea2ad147f60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ivan Horvat"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+38761234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dispatcher created successfully"
                    }
                }
            }
        },
        "/api/v1/dispatchers/{id}": {
            "get": {
                "tags": [
                    "Dispatchers"
                ],
                "summary": "Get a single dispatcher",
                "operationId": "98a5a5efc0634082c54e0c20ec47bc5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dispatcher data"
                    }
                }
            },
            "put": {
                "tags": [
                    "Dispatchers"
                ],
                "summary": "Update a dispatcher",
                "operationId": "ebe86ec7b2f16b1d7b30f4a879b304ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ivan Horvat Updated"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+38761234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dispatcher updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dispatchers"
                ],
                "summary": "Delete a dispatcher",
                "operationId": "536ffc7454ccb7340f61f88ec6fddf67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Dispatcher deleted successfully"
                    }
                }
            }
        },
        "/api/v1/locations": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get list of locations",
                "operationId": "42fc30e97773855c314df24aee584ce5",
                "responses": {
                    "200": {
                        "description": "List of locations"
                    }
                }
            },
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Create new location",
                "operationId": "32ee553d585d8b4295d1721b072d79af",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Konzum Mostar"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Mostar, BiH"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Location created successfully"
                    }
                }
            }
        },
        "/api/v1/locations/{id}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get a single location",
                "operationId": "3356f213a09d3820c976ab47617b23c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location data"
                    }
                }
            },
            "put": {
                "tags": [
                    "Locations"
                ],
                "summary": "Update a location",
                "operationId": "4b8dc7912804c3b07e6cb9c5396860c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Konzum Mostar Updated"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Mostar, BiH"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Locations"
                ],
                "summary": "Delete a location",
                "operationId": "fce9f4bc941c78a1dbecefc9529571d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Location deleted successfully"
                    }
                }
            }
        },
...
    },
...
}